refactor(ffa): extract pure garbage-queue helpers (plan §6A.2 first slice)
Move the two pure, test-covered garbage helpers out of the 4,728-line
ffa-p2p-game-state god-file into ffa/garbage-helpers.js — the first slice of
the future ffa/garbage-system.js, following the established net-diagnostics
pure-function extraction pattern.

- garbageBurstKey / drainAllLineBursts (no instance state) move verbatim; the
  five internal call sites use the imports; the class methods are removed.
- ffa-garbage-idempotent.test.js now imports the helpers directly (it already
  pinned them as pure — "does not use this"), strengthening the coverage.
- god-file 4,728 -> 4,708 lines (fitness ratchet lowered); the new module joins
  the @ts-check ratchet.

Behavior-preserving by construction (pure functions moved unchanged). Verified:
typecheck, ts-ratchet, lint:ci, check:boundaries, architecture-fitness, and the
full 2,396-test suite (incl. all 161 MP tests) green.
