+++ architecture-fitness.json
@@ -2,7 +2,7 @@
   "note": "Architecture ratchet baselines (Phase 3d) — shrink-only, see scripts/architecture-fitness-check.mjs. Each metric cites its plan section/ADR there. Never raise a baseline silently: a raise requires its own commit naming the plan-mandated change (see 2026-07-10 reconciliation: +1 main / +2 LMM / +2 OnlineMP were the §4.4 viewport + §4.6 emit-helper import lines).",
   "capturedAt": "2026-07-14",
   "metrics": {
-    "lines:ffa-p2p-game-state": 4728,
+    "lines:ffa-p2p-game-state": 4709,
     "lines:main": 5817,
     "lines:OdysseyMode": 4924,
     "lines:LocalMultiplayerMode": 4205,
